+++ insight/urls.py
@@ -6,7 +6,7 @@ from insight.views import set_origin_code
 urlpatterns = patterns('',
     # tracking url
     url(
-        r'^(?P<prefix>[a-z])/(?P<code>[\w]+)/$',
+        r'^i/(?P<code>[\w]+)/$',
         set_origin_code,
         name='set-origin-code'
     ),