+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.1 (02-04-2013)
+----------------
+#. Send `origin_hit` signal on url hit. Other apps can use it for custom tracking.
+#. Can enable/disable registration tracking per origin.
+#. Add ability to record querystring parameters in tracking urls.
+
 0.0.1 (28-06-2012)
 ------------------
 #. First release